# src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
...
paths:
...
  /emp/all_days:
    get:
      tags:
        - employeeController
      summary: get total work days for employee
      operationId: getWorkDaysByEmployee
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: query
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int32
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/date/absence:
    get:
      tags:
        - employeeController
      summary: get employees in absence on specified date
      operationId: getEmployeesDateInAbsence
      produces:
        - '*/*'
      parameters:
        - name: date
          in: query
          description: date
          required: true
          type: string
          format: date
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/date/at_work:
    get:
      tags:
        - employeeController
      summary: get employees at work in specified date
      operationId: getEmployeesDateAtWork
      produces:
        - '*/*'
      parameters:
        - name: date
          in: query
          description: date
          required: true
          type: string
          format: date
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/date/on_sick_leave:
    get:
      tags:
        - employeeController
      summary: get employees on sick leave at specified date
      operationId: getEmployeesDateOnSickLeave
      produces:
        - '*/*'
      parameters:
        - name: date
          in: query
          description: date
          required: true
          type: string
          format: date
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/date/on_vacation:
    get:
      tags:
        - employeeController
      summary: get employees on vacation at specified date
      operationId: getEmployeesDateOnVacation
      produces:
        - '*/*'
      parameters:
        - name: date
          in: query
          description: date
          required: true
          type: string
          format: date
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/delete:
    delete:
      tags:
        - employeeController
      summary: delete employee
      operationId: deleteEmployee
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: query
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /emp/new:
    post:
      tags:
        - employeeController
      summary: insert new employee
      operationId: insertEmployee
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: employeeDto
          description: employeeDto
          required: true
          schema:
            $ref: '#/definitions/EmployeeDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmployeeDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/now/absence:
    get:
      tags:
        - employeeController
      summary: get employees in absence now
      operationId: getEmployeesNowInAbsence
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/now/at_work:
    get:
      tags:
        - employeeController
      summary: get employees at work now
      operationId: getEmployeesNowAtWork
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/now/on_sick_leave:
    get:
      tags:
        - employeeController
      summary: get employees on sick leave now
      operationId: getEmployeesNowOnSickLeave
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/now/on_vacation:
    get:
      tags:
        - employeeController
      summary: get employees on vacation now
      operationId: getEmployeesNowOnVacation
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /emp/update:
    put:
      tags:
        - employeeController
      summary: update employee
      operationId: updateEmployee
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: employeeDto
          description: employeeDto
          required: true
          schema:
            $ref: '#/definitions/EmployeeDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmployeeDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/emp/{employeeId}':
    get:
      tags:
        - employeeController
      summary: get employee by id
      operationId: getEmployee
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: path
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmployeeDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
definitions:
...
  EmployeeDto:
    type: object
    properties:
      dateIn:
        type: string
        format: date
      dateOut:
        type: string
        format: date
      departmentId:
        type: integer
        format: int64
      firstName:
        type: string
      id:
        type: integer
        format: int64
      lastName:
        type: string
      positionId:
        type: integer
        format: int64
    title: EmployeeDto
...